+++ src/data/ImporterSpecData.ts
@@ -1,7 +1,7 @@
 import {AnnotationFormatType} from "./enums/AnnotationFormatType";
 import {AnnotationImporter} from "../logic/import/AnnotationImporter";
-import {COCOImporter} from "../logic/import/coco/importer";
-import {YOLOImporter} from "../logic/import/yolo/importer";
+import {COCOImporter} from "../logic/import/coco/COCOImporter";
+import {YOLOImporter} from "../logic/import/yolo/YOLOImporter";
 
 export type ImporterSpecDataMap = { [s in AnnotationFormatType]: typeof AnnotationImporter; };
